+++ reporter/doc.go
@@ -1,2 +1,4 @@
-// Package reporter provides reporters as [model.System] implementations.
+// Package reporter provides reporters as System implementations ([github.com/mlange-42/arche-model/model.System]).
+// Reporters serve for handling data extracted by an [observer],
+// e.g. by displaying it or writing it to a file.
 package reporter