+++ observer/observer_table_test.go
@@ -1,0 +1,35 @@
+package observer_test
+
+import (
+	"github.com/mlange-42/arche-model/model"
+	"github.com/mlange-42/arche-model/observer"
+	"github.com/mlange-42/arche/ecs"
+)
+
+// Example observer, reporting a nonsense table.
+type TableObserver struct{}
+
+func (o *TableObserver) Initialize(w *ecs.World) {}
+
+func (o *TableObserver) Update(w *ecs.World) {}
+
+func (o *TableObserver) Header() []string {
+	return []string{"TotalEntities"}
+}
+
+func (o *TableObserver) Values(w *ecs.World) [][]float64 {
+	return [][]float64{
+		{1, 2, 3},
+		{1, 2, 3},
+		{1, 2, 3},
+		{1, 2, 3},
+	}
+}
+
+func ExampleTable() {
+	m := model.New()
+
+	var obs observer.Table = &TableObserver{}
+	_ = obs.Values(&m.World)
+	// Output:
+}