+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Documentation
 
-* Extend documentation on `Model`, `Systems` and `Observer` (#18) [TEST]
+* Extend documentation on `Model`, `Systems` and `Observer` (#18)
 
 ## [[v0.0.2]](https://github.com/mlange-42/arche-model/compare/v0.0.1...v0.0.2)